+++ WhistleWind.AbnormalSigils/Abilities/ReturnOne.cs
@@ -1,6 +1,8 @@
-﻿using DiskCardGame;
+﻿using Core.Helpers;
+using DiskCardGame;
 using Infiniscryption.Spells.Patchers;
 using InscryptionAPI.Card;
+using InscryptionAPI.Helpers.Extensions;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -17,12 +19,28 @@ namespace WhistleWind.AbnormalSigils
         public override Ability Ability => ability;
 
         public override bool RespondsToResolveOnBoard() => base.Card.Info.IsGlobalSpell();
-        public override bool RespondsToSlotTargetedForAttack(CardSlot slot, PlayableCard attacker) => base.Card.Info.IsTargetedSpell() && slot.IsPlayerSlot && slot.Card != null;
+        public override bool RespondsToSlotTargetedForAttack(CardSlot slot, PlayableCard attacker)
+        {
+            if (base.Card.Info.IsTargetedSpell() && slot.Card != null)
+            {
+                return base.Card.OriginatedFromQueue ? !slot.IsPlayerSlot : slot.IsPlayerSlot;
+            }
+            return false;
+        }
 
         public override IEnumerator OnSlotTargetedForAttack(CardSlot slot, PlayableCard attacker)
         {
+            if (base.Card.OpponentCard)
+            {
+                ViewManager.Instance.SwitchToView(View.OpponentQueue);
+                yield return TurnManager.Instance.Opponent.ReturnCardToQueue(slot.Card, 0.2f);
+                slot.Card.UnassignFromSlot();
+                slot.Card = null;
+                yield return new WaitForSeconds(0.4f);
+                yield break;
+            }
+
             CardInfo copy = slot.Card.Info.Clone() as CardInfo;
-            // includes status effects
             PlayableCardStatus status = new(slot.Card.Status);
             List<CardModificationInfo> tempMods = slot.Card.TemporaryMods;
             tempMods.Add(new()
@@ -35,6 +53,7 @@ namespace WhistleWind.AbnormalSigils
 
             List<SpecialCardBehaviour> behaviours = slot.Card.GetComponents<SpecialCardBehaviour>()?.ToList() ?? new();
             List<AbilityBehaviour> abilities = slot.Card.GetComponents<AbilityBehaviour>()?.ToList() ?? new();
+
             slot.Card.RemoveFromBoard(false);
             yield return HelperMethods.ChangeCurrentView(View.Default);
             yield return CardSpawner.Instance.SpawnCardToHand(copy, tempMods, 0.25f, (PlayableCard x) =>
@@ -46,28 +65,42 @@ namespace WhistleWind.AbnormalSigils
                     x.TriggerHandler.permanentlyAttachedBehaviours.Add(copy);
                 }
             });
+
             yield return new WaitForSeconds(0.2f);
         }
         public override IEnumerator OnResolveOnBoard()
         {
-            foreach (var slot in BoardManager.Instance.PlayerSlotsCopy.Where(x => x.Card != null))
+            if (base.Card.OriginatedFromQueue)
             {
-                CardInfo copy = slot.Card.Info.Clone() as CardInfo;
-                PlayableCardStatus status = new(slot.Card.Status);
-                List<CardModificationInfo> tempMods = slot.Card.TemporaryMods;
-                tempMods.Add(new()
+                ViewManager.Instance.SwitchToView(View.OpponentQueue);
+                foreach (PlayableCard card in BoardManager.Instance.GetCards(false))
                 {
-                    bloodCostAdjustment = -slot.Card.BloodCost(),
-                    bonesCostAdjustment = GetBonesCost(slot.Card) - slot.Card.BonesCost(),
-                    energyCostAdjustment = -slot.Card.EnergyCost,
-                    nullifyGemsCost = true
-                });
-
-                slot.Card.RemoveFromBoard(false);
-                yield return HelperMethods.ChangeCurrentView(View.Default, 0.1f, 0.1f);
-                yield return CardSpawner.Instance.SpawnCardToHand(copy, tempMods, 0.25f, (PlayableCard x) => { x.Status = status; });
-                yield return new WaitForSeconds(0.2f);
+                    yield return TurnManager.Instance.Opponent.ReturnCardToQueue(card, 0.2f);
+                    card.Slot.Card = null;
+                    card.UnassignFromSlot();
+                }
             }
+            else
+            {
+                foreach (PlayableCard card in BoardManager.Instance.GetCards(true))
+                {
+                    CardInfo copy = card.Info.Clone() as CardInfo;
+                    PlayableCardStatus status = new(card.Status);
+                    List<CardModificationInfo> tempMods = card.TemporaryMods;
+                    tempMods.Add(new()
+                    {
+                        bloodCostAdjustment = -card.BloodCost(),
+                        bonesCostAdjustment = GetBonesCost(card) - card.BonesCost(),
+                        energyCostAdjustment = -card.EnergyCost,
+                        nullifyGemsCost = true
+                    });
+
+                    card.RemoveFromBoard(false);
+                    yield return HelperMethods.ChangeCurrentView(View.Default, 0.1f, 0.1f);
+                    yield return CardSpawner.Instance.SpawnCardToHand(copy, tempMods, 0.25f, (PlayableCard x) => { x.Status = status; });
+                }
+            }
+            yield return new WaitForSeconds(0.2f);
         }
         private int GetBonesCost(PlayableCard card) => Mathf.Max(0, 2 - (TurnManager.Instance.TurnNumber - card.TurnPlayed));
     }